Removed R8 configuration and disabled code shrinking.
It is no longer possible to produce human-readable stack traces while optimizing with R8 without a mapping file.
This adds some complexity that is not present in the vanilla codebase.
Furthermore vanilla PD never used proguard/R8 anyway, and so including it arguably breaks the functional equivalence with the original PD repo.